🛠️ Refactor: Extract magic numbers and merge sparse directories
Extracted magic numbers for dimensions, paddings, timings, and thresholds in `VibeboardKeyboard.kt` and `MainActivity.kt` into `KeyboardDimensions`, `KeyboardTiming`, and `SetupScreenDimensions` objects.
Merged the sparse `ui` directory into its parent `br.tec.lew.vibeboard` to group by domain instead of file type.
Added `mise.toml` to define the CI pipeline and configured `AGENTS.md` with Operational Memory.

As described by Martin Fowler in *Refactoring* ("Replace Magic Number with Symbolic Constant") and Robert C. Martin in *Clean Code*, replacing literal values with clearly named constants improves readability, makes the code self-documenting, and centralizes configuration for easier maintenance.
