+++ AGENTS.md
@@ -9,7 +9,7 @@ It uses Jetpack Compose for the UI.
 
 - `app/src/main/java/br/tec/lew/vibeboard/MainActivity.kt` -> Main entry point and setup screen.
 - `app/src/main/java/br/tec/lew/vibeboard/VibeboardService.kt` -> Core InputMethodService containing the logic of the keyboard.
-- `app/src/main/java/br/tec/lew/vibeboard/ui/VibeboardKeyboard.kt` -> The Compose UI for the keyboard.
+- `app/src/main/java/br/tec/lew/vibeboard/VibeboardKeyboard.kt` -> The Compose UI for the keyboard.
 - `app/src/main/java/br/tec/lew/vibeboard/ErrorHandler.kt` -> Centralized error reporting function (`reportError`).
 - `mise.toml` -> Configures the `mise` environment, tools, and executable tasks like `ci` and `test`.
